+++ README.md
@@ -6,6 +6,14 @@ Data retrieved live from aviation weather services, parsed and displayed to the 
 
 Click on the Runways for a fun animation. 
  
+## Technology
+React and D3 frontend powered by a nodejs backend. Hosted on a Google Cloud App Engine.
+
+## Why?
+I am a private pilot and wanted an easier way to check the weather for an airport from my phone or computer.  
+This app visualizes the weather and allows anyone to search for an airport and see the most updated information. 
+
+
 ### API for accessing live Metar and TAF data:
 
 #### /api/newestMETAR/{airport ident}
@@ -94,9 +102,3 @@ Example: /api/newestTAFS/KSLN
             ]
         }
 
-## Technology
-React and D3 frontend powered by a nodejs backend. Hosted on a Google Cloud App Engine.
-
-## Why?
-I am a private pilot and wanted an easier way to check the weather for an airport from my phone or computer.  
-This app visualizes the weather and allows anyone to search for an airport and see the most updated information.
