+++ react-app/package.json
@@ -6,6 +6,7 @@
     "axios": ">=0.18.1",
     "d3": "^5.12.0",
     "mixin-deep": ">=1.3.2",
+    "rc-slider": "^8.7.1",
     "react": "^16.9.0",
     "react-dom": "^16.9.0",
     "react-responsive": "^8.0.1",
