+++ react-app/src/components/index.js
@@ -1,0 +1,7 @@
+import AirplaneIcon from './AirplaneIcon';
+import ButtonLink from './Button';
+
+export {
+    AirplaneIcon,
+    ButtonLink
+}
