+++ react-app/src/weathervis/graphs/HistoricalWind.jsx
