+++ react-app/src/weathervis/onemetar/visibility.jsx
