+++ react-app/src/weathervis/onemetar/wind.jsx
